requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
camelot-py==0.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.7
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
multidict==4.7.6
numpy==1.19.0
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.5
pdfminer.six==20200517
pycryptodome==3.9.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
uWSGI==2.0.19.1
yarl==1.4.2
